update ecs task runner

+++ bin/service.js
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+/* eslint-disable no-console */
 
 const fs = require('fs');
 const path = require('path');
@@ -7,7 +8,12 @@ const minimist = require('minimist');
 const createCliOptions = require('cliclopts');
 const rimraf = require('rimraf');
 
-const runService = require('../index');
+const {
+  runServiceFromActivity,
+  runServiceFromSQS,
+  runTask,
+  logMessage
+} = require('../index');
 
 const taskDir = path.join(process.cwd(), 'task');
 const workDir = path.join(process.cwd(), '.tmp-work');
@@ -37,17 +43,44 @@ const cliOptions = createCliOptions([
     name: 'activity-arn',
     abbr: 'a',
     alias: ['activityArn'],
+    default: null,
     help: 'the arn of the step function activity for this task'
   },
   {
+    name: 'sqs-url',
+    alias: ['sqsUrl'],
+    default: null,
+    help: 'the sqs url used for getting messages for this task'
+  },
+  {
+    name: 'lambda-input',
+    alias: ['lambdaInput'],
+    default: null,
+    help: 'the message input for this task'
+  },
+  {
     name: 'heartbeat',
-    abbr: 'h',
+    default: null,
     help: 'interval in milliseconds between sending heartbeat messages to the state machine. ' +
     'default is null, which disables the heartbeat'
+  },
+  {
+    name: 'help',
+    abbr: 'h',
+    help: 'show help',
+    boolean: true
   }
 ]);
 
 const argv = minimist(process.argv.slice(2), cliOptions.options());
+
+if (argv.help) {
+  console.log('Usage: command [options]');
+  cliOptions.print();
+  process.exit(0);
+}
+
+logMessage('info', 'Starting the cumulus-ecs-task runner ...');
 
 rimraf.sync(argv.taskDirectory);
 rimraf.sync(argv.workDirectory);
@@ -55,6 +88,26 @@ rimraf.sync(argv.workDirectory);
 fs.mkdirSync(argv.taskDirectory);
 fs.mkdirSync(argv.workDirectory);
 
-runService(argv);
+if (argv.activityArn) {
+  runServiceFromActivity(argv).catch(console.error);
+}
+else if (argv.sqsUrl) {
+  runServiceFromSQS(argv).catch(console.error);
+}
+else if (argv.lambdaInput) {
+  argv.lambdaInput = JSON.parse(argv.lambdaInput);
+  runTask(argv).catch(console.error);
+}
+else {
+  logMessage(
+    'error',
+    'You must provider one of the following options: activity-arn, sqs-url, lambda-input'
+  );
+  process.exit(1);
+}
 
-process.on('exit', () => rimraf.sync(argv.workDirectory));
+process.on('SIGINT', () => {
+  logMessage('info', 'Doing some cleanup work before quitting!');
+  rimraf.sync(argv.workDirectory);
+  process.exit();
+});
